requirements.txt
boto3==1.34.13
python-dotenv==1.0.1
flask==3.0.1
pydantic==2.9.2
gnews==0.4.0